--- pxc-framework-common/src/main/java/io/github/panxiaochao/common/plugin/datasource/po/AbstractColumn.java
@@ -1,11 +1,0 @@
-package io.github.panxiaochao.common.plugin.datasource.po;
-
-/**
- * {@code AbstractColumn}
- * <p> description: 数据库字段 抽象类
- *
- * @author Lypxc
- * @since 2023-04-21
- */
-public abstract class AbstractColumn implements IColumn {
-}
