+++ pxc-framework-database-metadata/pom.xml
@@ -4,18 +4,22 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <artifactId>pxc-framework-parent</artifactId>
         <groupId>io.github.panxiaochao</groupId>
+        <artifactId>pxc-framework-parent</artifactId>
         <version>2022-2.7-1.4-SNAPSHOT</version>
-        <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <artifactId>pxc-framework-common</artifactId>
+    <artifactId>pxc-framework-database-metadata</artifactId>
     <packaging>jar</packaging>
 
-    <description>Framework Common Configuration Jar</description>
+    <description>Database Metadata Configuration Jar</description>
 
     <dependencies>
+        <dependency>
+            <groupId>io.github.panxiaochao</groupId>
+            <artifactId>pxc-framework-common</artifactId>
+            <version>${project.parent.version}</version>
+        </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-web</artifactId>
@@ -33,4 +37,5 @@
             <optional>true</optional>
         </dependency>
     </dependencies>
+
 </project>
