+++ pxc-framework-database-metadata/src/main/java/io/github/panxiaochao/database/metadata/core/database/AbstractDataBaseQuery.java
@@ -1,0 +1,38 @@
+package io.github.panxiaochao.database.metadata.core.database;
+
+import io.github.panxiaochao.database.metadata.api.IDataBaseQuery;
+import io.github.panxiaochao.database.metadata.builder.DefaultDataSourceBuilder;
+import io.github.panxiaochao.database.metadata.builder.DefaultRulesBuilder;
+import io.github.panxiaochao.database.metadata.core.DefaultQuerySqlDecorator;
+import io.github.panxiaochao.database.metadata.core.wrapper.DatabaseMetaDataWrapper;
+import lombok.Getter;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+/**
+ * {@code AbstractDataBaseQuery}
+ * <p> description: 数据库 查询抽象类
+ *
+ * @author Lypxc
+ * @since 2023-04-14
+ */
+@Getter
+public abstract class AbstractDataBaseQuery implements IDataBaseQuery {
+
+    protected final Logger logger = LoggerFactory.getLogger(getClass());
+
+    protected final DefaultQuerySqlDecorator querySqlDecorator;
+
+    protected final DefaultDataSourceBuilder dataSourceBuilder;
+
+    protected final DefaultRulesBuilder rulesBuilder;
+
+    protected final DatabaseMetaDataWrapper databaseMetaDataWrapper;
+
+    protected AbstractDataBaseQuery(DefaultDataSourceBuilder dataSourceBuilder, DefaultRulesBuilder rulesBuilder) {
+        this.querySqlDecorator = new DefaultQuerySqlDecorator(dataSourceBuilder);
+        this.dataSourceBuilder = dataSourceBuilder;
+        this.rulesBuilder = rulesBuilder;
+        this.databaseMetaDataWrapper = new DatabaseMetaDataWrapper(dataSourceBuilder);
+    }
+}
