+++ pxc-framework-database-metadata/src/main/java/io/github/panxiaochao/database/metadata/po/AbstractColumn.java
@@ -1,0 +1,11 @@
+package io.github.panxiaochao.database.metadata.po;
+
+/**
+ * {@code AbstractColumn}
+ * <p> description: 数据库字段 抽象类
+ *
+ * @author Lypxc
+ * @since 2023-04-21
+ */
+public abstract class AbstractColumn implements IColumn {
+}
