+++ pxc-framework-db-document/pom.xml
@@ -13,7 +13,7 @@
     <artifactId>pxc-framework-db-document</artifactId>
     <packaging>jar</packaging>
 
-    <description>Db Document configuration jar</description>
+    <description>Db Document Configuration Jar</description>
 
     <properties>
         <screw-core.version>1.0.5</screw-core.version>
