+++ pxc-framework-ip2region/pom.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>io.github.panxiaochao</groupId>
+        <artifactId>pxc-framework-parent</artifactId>
+        <version>2022-2.7-1.4-SNAPSHOT</version>
+    </parent>
+
+    <artifactId>pxc-framework-ip2region</artifactId>
+    <packaging>jar</packaging>
+
+    <description>Ip2region Configuration Jar</description>
+
+</project>
