+++ pxc-framework-log4j2/pom.xml
@@ -13,7 +13,7 @@
     <artifactId>pxc-framework-log4j2</artifactId>
     <packaging>jar</packaging>
 
-    <description>Log4j2 configuration jar</description>
+    <description>Log4j2 Configuration Jar</description>
 
     <properties>
         <log4j2.version>2.19.0</log4j2.version>
