--- pxc-framework-mybatis-flex/pom.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>io.github.panxiaochao</groupId>
-        <artifactId>pxc-framework-parent</artifactId>
-        <version>2022-2.7-1.4-SNAPSHOT</version>
-    </parent>
-    <artifactId>pxc-framework-mybatis-flex</artifactId>
-    <packaging>jar</packaging>
-    <description>Mybatis-Flex configuration jar</description>
-
-    <properties>
-        <mybatis-flex.version>1.0.6</mybatis-flex.version>
-    </properties>
-
-    <dependencies>
-        <dependency>
-            <groupId>com.mybatis-flex</groupId>
-            <artifactId>mybatis-flex-spring-boot-starter</artifactId>
-            <version>${mybatis-flex.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-configuration-processor</artifactId>
-            <optional>true</optional>
-            <scope>provided</scope>
-        </dependency>
-    </dependencies>
-</project>
